+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.10.0
  - feat: add `NesIcons.user`
  - feat: add `NesIcons.chest`
+ - feat: add `NesIcons.lamp`
 
 # 0.9.0
  - feat: add `NesIcons.expand`